Added Segmentation Base Features, Signal Processing Features, Model hyperparameters, NaN propagation of missing values, changing underlying feature structure to keep names and values together.

+++ src/pose_estimation/pose_est_v2.py
@@ -39,7 +39,7 @@ class PoseEstimationV2(PoseEstimation):
             pose_grp = pose_h5['poseest']
 
             # load contents
-            self._points = pose_grp['points'][:]
+            self._points = pose_grp['points'][:].astype(np.float64)
             self._point_mask = np.zeros(self._points.shape[:-1], dtype=np.uint16)
             self._point_mask[:] = pose_grp['confidence'][:] > 0.3
